Починил проф скиллы

+++ src/methodist/StartPage/SkillsForm/ProfessionalSkillsList/ProfessionalSkill/ProfessionalSkill.jsx
@@ -5,15 +5,16 @@ import { ReactComponent as IcDropdownTop } from "assets/icons/ic_dropdown-indica
 import { ReactComponent as IcAlert } from "assets/icons/ic_alert.svg";
 import Button from "components/Form/Button/Button";
 import Modal from "components/Form/Modal/Modal";
-import AdditionalSkills from "../AdditionalSkill/AdditionalSkill";
+import ProfessionalSubskillsMap from "./SubskillsMap/ProfessionalSubskillsMap";
+import AdditionalSubskillsMap from "./SubskillsMap/AdditionalSubskillsMap";
 
 import store from "store/store";
 
 import "./ProfessionalSkill.css";
-import ProfessionalSubskillsMap from "./ProfessionalSubskillsMap/ProfessionalSubskillsMap";
 
 const ProfessionalSkill = observer(({ profSkill, isAdditional, isDemo }, key) => {
   const professionalSkillsList = store.skills.professionalSkillsList
+  const btnLabel = isAdditional ? "добавить" : "удалить"
 
   const [isOpen, setIsOpen] = useState(false);
 
@@ -36,46 +37,48 @@ const ProfessionalSkill = observer(({ profSkill, isAdditional, isDemo }, key) =>
     if (profSkill.additionalIsActive) professionalSkillsList.setAdditionalIsActive(profSkill.id);
   };
 
+  if (isDemo) return (
+    <li className="professional-skills__item" key={key}>
+      <div className="professional-skills__dropdown-indicator">
+        <span className="professional-skills__label">{profSkill.label}</span>
+      </div>
+    </li>
+  )
   return (
     <li className="professional-skills__item" key={key}>
       <div
         className="professional-skills__accordion-header"
-        onClick={() => {
-          professionalSkillsList.setIsActive(profSkill.id);
-        }}
+        onClick={() => professionalSkillsList.setIsActive(profSkill.id)}
       >
         <div
           className="professional-skills__dropdown-indicator"
           aria-expanded={!profSkill.isActive}
         >
-          <IcDropdownTop />
+          {!isAdditional && !isDemo && (<IcDropdownTop />)}
         </div>
         <span className="professional-skills__label">{profSkill.label}</span>
-        {!isDemo && <Button
+
+        {!isDemo && (<Button
           text={isAdditional ? "Добавить" : "Удалить"}
-          theme={isAdditional ? "default" : "outlined"}
-          width="100px"
-          height="25px"
-          fontSize="16px"
-          borderRadius="5px"
+          theme={isAdditional ? "filled" : "outlined"}
+          size="small"
           onClick={(event) => deleteBtnOnClick(event)}
+        />)}
+
+        <Modal
+          isOpen={isOpen}
+          header={<IcAlert />}
+          label={`Вы уверены, что хотите ${btnLabel} компетенцию "${profSkill.label}"`}
+          onAcceptClick={() => acceptBtnOnClick()}
+          onCancelClick={() => cancelBtnOnClick()}
         />
-        }
-        {isOpen && (
-          <Modal
-            header={<IcAlert />}
-            label={isAdditional ? "Вы уверены, что хотите добавить компетенцию \"" + profSkill.label + "\"?" : "Вы уверены, что хотите удалить компетенцию \"" + profSkill.label + "\"?"}
-            onAcceptClick={() => acceptBtnOnClick()}
-            onCancelClick={() => cancelBtnOnClick()}
-          />
-        )}
       </div>
       <div
         className="professional-skills__accordion-body"
         aria-expanded={!profSkill.isActive}
       >
-        <ProfessionalSubskillsMap profSkill={profSkill} type="Default"/>
-        <AdditionalSkills profSkill={profSkill} type="Additional" />
+        <ProfessionalSubskillsMap profSkill={profSkill} isAdditional={false} />
+        <AdditionalSubskillsMap profSkill={profSkill} isAdditional={true} />
       </div>
     </li>
   );
